added float for area to handle decimal numbers aswell

+++ Country_data.py
@@ -3,12 +3,12 @@ from Countries_info import countries
 
 class Country:
     def __init__(self, name, population, capital, area):
-        if int(population) < 0 or int(area) < 0:
+        if int(population) < 0 or float(area) < 0:
             raise ValueError("Population and area must be non-negative")
         self.name = name
         self.population = int(population)
         self.capital = capital
-        self.area = int(area)
+        self.area = float(area)
         self.subdivisions = []  # A list to store subdivisions
 
     def __str__(self):  # Returns a string representation of the Country
@@ -45,12 +45,12 @@ class Country:
 
 class Subdivision:
     def __init__(self, name, population, capital, area):
-        if int(population) < 0 or int(area) < 0:
+        if int(population) < 0 or float(area) < 0:
             raise ValueError("Population and area must be non-negative")
         self.name = name
         self.population = int(population)
         self.capital = capital
-        self.area = int(area)
+        self.area = float(area)
         self.subdivisions = []  # This stores the smaller subdivisions
 
     def __str__(self):  # Returns a string representation of the subdivisions
